+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 CHANGELOG
 =========
 
-[Next release](https://github.com/rebing/graphql-laravel/compare/v1.23.0...v1)
-------------
+[Next release](https://github.com/rebing/graphql-laravel/compare/v1.24.0...v1)
+--------------
+
+2019-06-10, v1.24.0
+-------------------
 ### Changed
 - Prefix named GraphiQL routes with `graphql.` for compatibility with Folklore [\#360](https://github.com/rebing/graphql-laravel/pull/360)